Fix the IsItemDeletionInProgressMutator test
