+++ ui-components/tests/test-index.js
@@ -29,5 +29,5 @@ import * as isRenameItemLoadingMutator from './state/mutators/is-rename-item-loa
 import * as renameItemLoadingErrorMutator from './state/mutators/rename-item-loading-error-mutator';
 import * as listItem from './components/list-item';
 import * as removeItemAction from './state/actions/remove-item-action';
-import * as isDeleteItemLoadingMutator from './state/mutators/is-delete-item-loading-mutator';
+import * as isDeleteItemLoadingMutator from './state/mutators/is-item-deletion-in-progress-mutator';
 import * as deleteItemLoadingErrorMutator from './state/mutators/delete-item-loading-error-mutator';
